Catalog detail: move the presentational widget tail to lib/widgets/detail
Lifts the 18 private presentational widgets below the State in
`lib/screens/catalog_item_detail_screen.dart` (origin lines 1767-3388) into
nine files under `lib/widgets/detail/`, one per family:

  catalog_detail_reveal.dart          _Reveal
  catalog_detail_backdrop.dart        _Backdrop
  catalog_detail_badges.dart          _CertBadge _GenreChip _MetacriticBadge
                                      _CastAvatar
  catalog_detail_rec_card.dart        _RecCard
  catalog_detail_description.dart     _Description _ReadMoreToggle
  catalog_detail_action_row.dart      _ActionRow (+ the file-local
                                      _kNetflixRed)
  catalog_detail_quick_actions.dart   _QuickActions _QuickAction
                                      _MdblistQuickActions
                                      _SimklQuickActions _SimklQuickAction
  catalog_detail_primary_button.dart  _PrimaryButton
  catalog_detail_glass.dart           _GlassCard _GlassIconButton

Each becomes public with a `CatalogDetail` prefix. The bare `Detail` prefix
was not available: `lib/widgets/detail/detail_identity.dart` already ships a
different `DetailActionRow` and `DetailPrimaryButton` for the merged detail
page, and two same-named widgets in one directory would be a trap.

Move only. Apart from the renames the diff against the origin is 79 lines,
all of them either the `super.key` a public widget constructor needs or the
`dart format` rewrap the longer names force. No behaviour change, no
forwarders, no `part of` / `extension on`, and zero references to the host's
private members: everything already travelled by constructor.

The host loses its now-unused `movie_watched_badge.dart` and `tv_keys.dart`
imports and drops from 3388 to 1771 lines.

Pinned by test/catalog_detail_widgets_origin_pin_test.dart, which passes
unchanged before and after. Mutation-checked on both sides: 18 one-line
mutations (one per moved class) fail 16 of its 26 tests.
