+++ lib/widgets/detail/catalog_detail_glass.dart
@@ -1,0 +1,79 @@
+/// The glass surfaces on `CatalogItemDetailScreen`: the info/details card and
+/// the round back button over the backdrop.
+library;
+
+import 'package:flutter/material.dart';
+
+// ── Glass card ────────────────────────────────────────────────────────────
+
+class CatalogDetailGlassCard extends StatelessWidget {
+  final List<Widget> children;
+  const CatalogDetailGlassCard({super.key, required this.children});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(18),
+      decoration: BoxDecoration(
+        color: Colors.white.withValues(alpha: 0.05),
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: Colors.white.withValues(alpha: 0.08),
+          width: 0.5,
+        ),
+        boxShadow: const [
+          BoxShadow(color: Color(0x22000000), blurRadius: 20, spreadRadius: 2),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        mainAxisSize: MainAxisSize.min,
+        children: children,
+      ),
+    );
+  }
+}
+
+// ── Glass icon button (back) ────────────────────────────────────────────────
+
+class CatalogDetailGlassIconButton extends StatelessWidget {
+  final IconData icon;
+  final VoidCallback onTap;
+  const CatalogDetailGlassIconButton({
+    super.key,
+    required this.icon,
+    required this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Material(
+      color: Colors.transparent,
+      child: InkWell(
+        onTap: onTap,
+        borderRadius: BorderRadius.circular(999),
+        child: Container(
+          width: 42,
+          height: 42,
+          decoration: BoxDecoration(
+            color: Colors.black.withValues(alpha: 0.55),
+            shape: BoxShape.circle,
+            border: Border.all(
+              color: Colors.white.withValues(alpha: 0.18),
+              width: 0.5,
+            ),
+            boxShadow: const [
+              BoxShadow(
+                color: Color(0x44000000),
+                blurRadius: 12,
+                spreadRadius: 1,
+              ),
+            ],
+          ),
+          child: Icon(icon, color: Colors.white, size: 22),
+        ),
+      ),
+    );
+  }
+}
